ada: Disable inlining in potentially unevaluated contexts
Instead of explicitly disabling inlining in quantified expressions,
(which happen to be only preanalysed) and then disabling inlining in
potentially unevaluated contexts that are fully analysed (which happen
to include quantified expressions), we now simply disable inlining in
all potentially unevaluated contexts, regardless of the full analysis
mode.

This also disables inlining in iterated component associations, which
can be both preanalysed or fully analysed depending on their expression,
but nevertheless are potentially unevaluated.

gcc/ada/

	* sem_res.adb (Resolve_Call): Replace early call to
	In_Quantified_Expression with a call to Is_Potentially_Unevaluated that
	was only done when Full_Analysis is true.

Tested on x86_64-pc-linux-gnu, committed on master.