ada: Recognize iterated_component_association as repeatedly evaluated
As iterated_component_association is an array_component_association
(because of a grammar rule Ada 2022 RM 4.3.3(5/5)), its expression is
repeatedly evaluated (because of Ada 2022 RM 6.1.1(22.14/5)).

With this patch we will now get errors for both conjuncts in this code,
which have semantically equivalent array aggregates that use an ordinary
component association and iterated component association.

  procedure Iter (S : String)
    with Post => String'(for J in 1 .. 3 => S (S'First)'Old) =
                 String'(         1 .. 3 => S (S'First)'Old);

gcc/ada/

	* sem_util.adb (Is_Repeatedly_Evaluated): Recognize iterated component
	association as repeatedly evaluated.

Tested on x86_64-pc-linux-gnu, committed on master.