vect: Vectorize via libfuncs
This patch allows vectorization when operators are available as
libfuncs, rather that only as insns.

This will be useful for amdgcn where we plan to vectorize loops that
contain integer division or modulus, but don't want to generate inline
instructions for the division algorithm every time.

The change should have not affect architectures that do not define
vector-mode libfuncs.

OK for mainline?

Andrew
vect: vectorize via libfuncs

This patch allows vectorization when the libfuncs are defined.

gcc/ChangeLog:

	* tree-vect-generic.cc: Include optabs-libfuncs.h.
	(get_compute_type): Check optab_libfunc.
	* tree-vect-stmts.cc: Include optabs-libfuncs.h.
	(vectorizable_operation): Check optab_libfunc.