ada: Fix missing error on function call returning incomplete view
Testing for the presence of Non_Limited_View is not sufficient to detect
whether the nonlimited view has been analyzed because Build_Limited_Views
always sets the field on the limited view.  Instead the discriminant is
whether this nonlimited view is itself an incomplete type.

gcc/ada/

	* sem_ch4.adb (Analyze_Call): Adjust the test to detect the presence
	of an incomplete view of a type on a function call.

Tested on x86_64-pc-linux-gnu, committed on master.