ada: Fix missing finalization for aggregates nested in conditional expressions
The finalization actions for the components of the aggregates are blocked
by Expand_Ctrl_Function_Call, which sets Is_Ignored_Transient on all the
temporaries generated from within a conditional expression whatever the
intermediate constructs.  Now aggregates and their expansion in the form
of block and loop statements are "impenetrable" as far as temporaries are
concerned, i.e. the lifetime of temporaries generated within them does
not extend beyond them, so their finalization must not be blocked there.

gcc/ada/

	* exp_util.ads (Within_Case_Or_If_Expression): Adjust description.
	* exp_util.adb (Find_Hook_Context): Stop the search for the topmost
	conditional expression, if within one, at contexts where temporaries
	may be contained.
	(Within_Case_Or_If_Expression): Return false upon first encoutering
	contexts where temporaries may be contained.

Tested on x86_64-pc-linux-gnu, committed on master.