c: add name hints to c_parser_declspecs [PR107583]
PR c/107583 notes that we weren't issuing a hint for

  struct foo {
    time_t mytime; /* missing <time.h> include should trigger fixit */
  };

in the C frontend.

The root cause is that one of the "unknown type name" diagnostics
was missing logic to emit hints, which this patch fixes.

Successfully bootstrapped & regrtested on x86_64-pc-linux-gnu.
Pushed to trunk as r14-1876-g57446d1bc9757e.

gcc/c/ChangeLog:
	PR c/107583
	* c-parser.cc (c_parser_declspecs): Add hints to "unknown type
	name" error.

gcc/testsuite/ChangeLog:
	PR c/107583
	* c-c++-common/spellcheck-pr107583.c: New test.