Optimize std::max early
Hi,
we currently produce very bad code on loops using std::vector as a stack, since
we fail to inline push_back which in turn prevents SRA and we fail to optimize
out some store-to-load pairs (PR109849).

I looked into why this function is not inlined and it is inlined by clang.  We
currently estimate it to 66 instructions and inline limits are 15 at -O2 and 30
at -O3.  Clang has similar estimate, but still decides to inline at -O2.

I looked into reason why the body is so large and one problem I spotted is the
way std::max is implemented by taking and returning reference to the values.

  const T& max( const T& a, const T& b );

This makes it necessary to store the values to memory and load them later
(Max is used by code computing new size of vector on resize.)
Two stores, conditional and load accounts as 8 instructions, while
MAX_EXPR as 1 and has a lot better chance to fold with the surrounding
code.

We optimize this to MAX_EXPR, but only during late optimizations.  I think this
is a common enough coding pattern and we ought to make this transparent to
early opts and IPA.  The following is easist fix that simply adds phiprop pass
that turns the PHI of address values into PHI of values so later FRE can
propagate values across memory, phiopt discover the MAX_EXPR pattern and DSE
remove the memory stores.

Bootstrapped/regtested x86_64-linux, does this look resonable thing to do?

Looking into how expensive the pass is, I think it is very cheap, except
that it computes postdominator and updates ssa even if no patterns
are matched.  I will send patch to avoid that.

gcc/ChangeLog:

	PR tree-optimization/109811
	PR tree-optimization/109849
	* passes.def: Add phiprop to early optimization passes.
	* tree-ssa-phiprop.cc: Allow clonning.

gcc/testsuite/ChangeLog:

	PR tree-optimization/109811
	PR tree-optimization/109849
	* gcc.dg/tree-ssa/phiprop-1.c: New test.