RISC-V: Fix out of range memory access of machine mode table
We extend the machine mode from 8 to 16 bits already. But there still
one placing missing from the tree-streamer. It has one hard coded array
for the machine code like size 256.

In the lto pass, we memset the array by MAX_MACHINE_MODE count but the
value of the MAX_MACHINE_MODE will grow as more and more modes are added.
While the machine mode array in tree-streamer still leave 256 as is.

Then, when the MAX_MACHINE_MODE is greater than 256, the memset of
lto_output_init_mode_table will touch the memory out of range unexpected.

This patch would like to take the MAX_MACHINE_MODE as the size of the
array in tree-streamer, to make sure there is no potential unexpected
memory access in future.

Signed-off-by: Pan Li <pan2.li@intel.com>

gcc/ChangeLog:

	* lto-streamer-in.cc (lto_input_mode_table): Use
	MAX_MACHINE_MODE for memory allocation.
	* tree-streamer.cc: Use MAX_MACHINE_MODE for array size.
	* tree-streamer.h (streamer_mode_table): Ditto.
	(bp_pack_machine_mode): Ditto.
	(bp_unpack_machine_mode): Ditto.